pyproject.toml
[tool.poetry]
name = "graduate-work"
version = "0.1.0"
description = ""
authors = ["VladmimirShtefan <shtefanvladimir14@gmail.com>"]
readme = "README.md"
packages = [{include = "graduate_work"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
pydantic = "^1.10.4"
requests = "^2.28.2"
logging = "^0.4.9.6"
django-redis = "^5.2.0"
redis = "^4.4.2"


[tool.poetry.group.dev.dependencies]
ansible-vault = "^2.1.0"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
